pyproject.toml
[project]
name = "todo-list-django-rest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8",
    "djangorestframework>=3.16.1",
]
[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "black>=25.1.0",
    "isort>=6.0.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.5",
    "snowflake>=1.6.0",
]

[tool.ruff]
line-length = 79
target-version = "py311"
exclude = ["**/migrations/**"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle (inclui espaçamento, espaços desnecessários)
    "F",    # pyflakes
    "I",    # isort (ordem de imports)
    "Q",    # flake8-quotes (aspas)
]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"
multiline-quotes = "double"
docstring-quotes = "double"
avoid-escape = true

# Ordem alfabética de imports
[tool.isort]
profile = "black"
line_length = 79
force_alphabetical_sort_within_sections = true
force_sort_within_sections = true

# Black para garantir compatibilidade de formatação
[tool.black]
line-length = 79
target-version = ["py311"]